# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "kuSystem Backend API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local dev"
    }
  ],
  "components": {
    "schemas": {
...
      "QuoteItem": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          },
          "discount": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          },
          "taxRate": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 1
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "quantity",
          "unitPrice"
        ]
      },
...
      "AdditionalCharge": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1
          },
          "amount": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "type",
          "amount"
        ]
      },
      "Quote": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "draft",
              "sent",
              "accepted",
              "rejected",
              "expired"
            ]
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "minLength": 1
          },
          "branchId": {
            "type": "string",
            "nullable": true
          },
          "branchName": {
            "type": "string",
            "nullable": true
          },
          "issueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "maxLength": 10
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "printNotes": {
            "type": "boolean",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuoteItem"
            }
          },
          "additionalCharges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdditionalCharge"
            }
          },
          "id": {
            "type": "string"
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "subtotal": {
            "type": "number",
            "nullable": true
          },
          "taxTotal": {
            "type": "number",
            "nullable": true
          },
          "discountTotal": {
            "type": "number",
            "nullable": true
          },
          "total": {
            "type": "number",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "customerName",
          "items",
          "id"
        ]
      },
      "QuoteInput": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "draft",
              "sent",
              "accepted",
              "rejected",
              "expired"
            ]
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "minLength": 1
          },
          "branchId": {
            "type": "string",
            "nullable": true
          },
          "branchName": {
            "type": "string",
            "nullable": true
          },
          "issueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "maxLength": 10
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "printNotes": {
            "type": "boolean",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "nullable": true
                },
                "description": {
                  "type": "string",
                  "minLength": 1
                },
                "quantity": {
                  "type": "number",
                  "minimum": 0
                },
                "unitPrice": {
                  "type": "number",
                  "minimum": 0
                },
                "discount": {
                  "type": "number",
                  "nullable": true,
                  "minimum": 0
                },
                "taxRate": {
                  "type": "number",
                  "nullable": true,
                  "minimum": 0,
                  "maximum": 1
                }
              },
              "required": [
                "description",
                "quantity",
                "unitPrice"
              ]
            },
            "default": []
          },
          "additionalCharges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "minLength": 1
                },
                "amount": {
                  "type": "number",
                  "minimum": 0
                }
              },
              "required": [
                "type",
                "amount"
              ]
            },
            "default": []
          }
        },
        "required": [
          "customerName"
        ]
      },
      "QuoteInputPartial": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "draft",
              "sent",
              "accepted",
              "rejected",
              "expired"
            ]
          },
          "customerId": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "minLength": 1
          },
          "branchId": {
            "type": "string",
            "nullable": true
          },
          "branchName": {
            "type": "string",
            "nullable": true
          },
          "issueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "currency": {
            "type": "string",
            "nullable": true,
            "maxLength": 10
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "printNotes": {
            "type": "boolean",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "nullable": true
                },
                "description": {
                  "type": "string",
                  "minLength": 1
                },
                "quantity": {
                  "type": "number",
                  "minimum": 0
                },
                "unitPrice": {
                  "type": "number",
                  "minimum": 0
                },
                "discount": {
                  "type": "number",
                  "nullable": true,
                  "minimum": 0
                },
                "taxRate": {
                  "type": "number",
                  "nullable": true,
                  "minimum": 0,
                  "maximum": 1
                }
              },
              "required": [
                "description",
                "quantity",
                "unitPrice"
              ]
            },
            "default": []
          },
          "additionalCharges": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "minLength": 1
                },
                "amount": {
                  "type": "number",
                  "minimum": 0
                }
              },
              "required": [
                "type",
                "amount"
              ]
            },
            "default": []
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/quotes": {
      "get": {
        "summary": "List quotes",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer"
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of quotes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Quote"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "pageSize": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "pageSize"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create quote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{id}": {
      "get": {
        "summary": "Get quote",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update quote",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInputPartial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete quote",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}